requirements.txt
# VectraHub GPU Server - Python Dependencies

# Core dependencies
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0

# Image processing
Pillow>=9.5.0
opencv-python-headless>=4.8.0
scikit-image>=0.21.0
scipy>=1.11.0
cairosvg>=2.7.0

# Deep Learning / Transformers
transformers>=4.36.0
accelerate>=0.25.0
peft>=0.7.0
huggingface-hub>=0.19.0
timm>=0.9.12

# Real-ESRGAN and BasicSR
basicsr>=1.4.2
realesrgan>=0.3.0
gfpgan>=1.3.8
facexlib>=0.3.0
filterpy>=1.4.5
numba>=0.58.0

# VTracer dependencies
vtracer>=0.6.11
pymatting>=1.1.8

# SVG processing
svglib>=1.5.1
svgpathtools>=1.6.1
lxml>=4.9.3
svgwrite>=1.4.3

# API / Web
requests>=2.31.0
aiohttp>=3.9.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Utilities
pydantic>=2.5.0
python-dotenv>=1.0.0
tqdm>=4.66.0
colorama>=0.4.6
psutil>=5.9.6

# OmniSVG dependencies (from their repo)
einops>=0.7.0
sentencepiece>=0.1.99
decord>=0.6.0
qwen-vl-utils>=0.0.2
